issue #935 extract tau-gateway crate
