+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
   "crates/tau-core",
   "crates/tau-ai",
   "crates/tau-agent-core",
+  "crates/tau-gateway",
   "crates/tau-runtime",
   "crates/tau-tui",
   "crates/tau-coding-agent",
