+++ crates/tau-coding-agent/Cargo.toml
@@ -17,6 +17,7 @@ ed25519-dalek.workspace = true
 futures-util.workspace = true
 hmac.workspace = true
 tau-core = { path = "../tau-core" }
+tau-gateway = { path = "../tau-gateway" }
 tau-runtime = { path = "../tau-runtime" }
 tau-agent-core = { path = "../tau-agent-core" }
 tau-ai = { path = "../tau-ai" }
