+++ crates/tau-gateway/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "tau-gateway"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+tau-core = { path = "../tau-core" }
+tau-runtime = { path = "../tau-runtime" }
+tokio.workspace = true
+
+[dev-dependencies]
+tempfile = "3"
