+++ crates/tau-gateway/src/lib.rs
@@ -1,0 +1,9 @@
+pub mod gateway_contract;
+pub mod gateway_runtime;
+pub mod gateway_ws_protocol;
+pub mod remote_profile;
+
+pub use gateway_contract::*;
+pub use gateway_runtime::*;
+pub use gateway_ws_protocol::*;
+pub use remote_profile::*;
