+++ crates/tau-gateway/testdata/gateway-contract/invalid-error-code.json
