+++ crates/tau-gateway/testdata/gateway-contract/mixed-outcomes.json
